Clean up bukkit listeners by splitting them up
